requirements.txt
chromadb==1.0.12
filetype==1.2.0
langchain==0.3.27
langchain-core==0.3.76
langchain-google-genai==2.1.10
langchain-groq==0.3.8
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
numpy==2.3.3
python-dotenv==1.1.1
PyYAML==6.0.2
sentence-transformers==5.1.0
tiktoken==0.11.0
typing-inspection==0.4.1
typing_extensions==4.15.0
PyPDF2==3.0.1
streamlit==1.50.0
